+++ src/main/java/it/tredi/msa/configuration/MSAConfigurationReader.java
@@ -13,7 +13,9 @@ public class MSAConfigurationReader {
 	private final static String MAILBOXMANAGERS_POOLSIZE_PROPERTY = "mailboxmanagers.poolsize";
 	private final static String MAILBOXMANAGERS_REFRESHTIME_PROPERTY = "mailboxmanagers.refresh.time";
 	private final static String MAILBOXMANAGERS_HOT_RELOADING_PROPERTY = "mailboxmanagers.hot-reloading";
-	public final static String MAILBOXMANAGERS_ALLOW_EMAIL_DUPLICATES_PROPERTY = "mailboxmanagers.allow-email-duplicates";
+	
+	// mbernardini 26/02/2019 : forzata l'univocita' degli indirizzi delle caselle di posta gestite da msa
+	//public final static String MAILBOXMANAGERS_ALLOW_EMAIL_DUPLICATES_PROPERTY = "mailboxmanagers.allow-email-duplicates";
 
 	public MSAConfiguration read() throws Exception {
 		PropertiesReader propertiesReader = new PropertiesReader(PROPERTIES_FILENAME);
@@ -27,7 +29,11 @@ public class MSAConfigurationReader {
 		msaConfiguration.setMailboxManagersDelay(propertiesReader.getIntProperty(MAILBOXMANAGERS_DELAY_PROPERTY, 600));
 		msaConfiguration.setMailboxManagersPoolsize(propertiesReader.getIntProperty(MAILBOXMANAGERS_POOLSIZE_PROPERTY, 1));
 		msaConfiguration.setMailboxManagersRefreshTime(propertiesReader.getIntProperty(MAILBOXMANAGERS_REFRESHTIME_PROPERTY, 1));
-		msaConfiguration.setAllowEmailDuplicates(propertiesReader.getBooleanProperty(MAILBOXMANAGERS_ALLOW_EMAIL_DUPLICATES_PROPERTY, false));
+		
+		// mbernardini 26/02/2019 : forzata l'univocita' degli indirizzi delle caselle di posta gestite da msa
+		//msaConfiguration.setAllowEmailDuplicates(propertiesReader.getBooleanProperty(MAILBOXMANAGERS_ALLOW_EMAIL_DUPLICATES_PROPERTY, false));
+		msaConfiguration.setAllowEmailDuplicates(false);
+		
 		msaConfiguration.setMailboxManagersHotReloading(propertiesReader.getBooleanProperty(MAILBOXMANAGERS_HOT_RELOADING_PROPERTY, false));
 		
 		//AuditWriter configuration